A working experiment with a moving snake.
This experiment does not include the adding of body parts, but simply
contains a classless simplification of the game snake. I just needed
something to get some code on the board.

I decided to code up my snake such that the squares the snake occupies
control the relative velocity of the snake.