Added some intelligence to the apple placement.
There is still a bug where when my snake is adjacent to the apple,
with my velocity facing away from the apple, and then turning at a 90
degree angle to grab the apple in the next frame. that seems to be the
instance where this fails because my edge detection is wrong there.